% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{a_check}
\alias{a_check}
\title{a_check}
\usage{
a_check(imageId, goldStandard)
}
\arguments{
\item{imageId}{the image ID i.e. filename without extension}

\item{goldStandard}{a list of annotated markers for the corresponding image}
}
\value{
An object that represents an A-Frame Event.
}
\description{
Check if annotations are correct or incorrect
}
\details{
Display the annotated marker ring entities as green if correct or red if incorrect
}
