% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scene.R
\name{a_kill_all_scenes}
\alias{a_kill_all_scenes}
\title{a_kill_all_scenes}
\usage{
a_kill_all_scenes()
}
\value{
nothing.
}
\description{
Kill all running A-Frame Scenes
}
\details{
Frequently in development you may accidentally loose a handle to a running
scene, meaning it will block a port you wish to run a new scene on. This
kills all running scenes to free up all ports being used to serve by r2vr.
}
