% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{a_label}
\alias{a_label}
\title{a_label}
\usage{
a_label(text, color = "#000000", font = "roboto", ...)
}
\arguments{
\item{text}{the label text.}

\item{color}{the label text colour.}

\item{font}{the font to be used for the label text. Either the name of a
built-in font, listed here:
https://aframe.io/docs/0.8.0/components/text.html#stock-fonts, or a URL
for an sdf .fnt file.}

\item{...}{additional arguments to be passed to the A-Frame text entity,
see: https://aframe.io/docs/0.8.0/primitives/a-text.html.}
}
\value{
an A-Frame entity object.
}
\description{
Create a text label, visible in VR.
}
\details{
This function returns an entity that will render a 2D text label in VR. It
is wrapper for an A-Frame entity with the text component attached, see:
https://aframe.io/docs/0.8.0/components/text.html. Additional component
configuration arguments listed in the A-Frame documentation can be supplied
in the `...` arguments.

Use standard entity components such as `position`, `rotation`, and `scale`
in `...` to place the label.
}
