% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{a_remove_entity}
\alias{a_remove_entity}
\title{a_remove_entity}
\usage{
a_remove_entity(id)
}
\arguments{
\item{id}{id of the entity to be removed.}
}
\value{
An object that represents an A-Frame Event.
}
\description{
Remove an A-Frame entity from the scene
}
\details{
Remove the entity identified by 'id' from the scene.
}
