% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_question_scene.R
\name{binary_question_scene}
\alias{binary_question_scene}
\title{Create a binary question scene}
\usage{
binary_question_scene(
  the_question,
  answer_1,
  answer_2,
  img_paths,
  IPv4_ADDRESS,
  animal_class,
  message_height = 1.5
)
}
\arguments{
\item{the_question}{Question message to display}

\item{answer_1}{Binary answer for \code{1} (Yes)}

\item{answer_2}{Binary answer for \code{0} (No)}

\item{img_paths}{Character string of image paths from current working directory}

\item{IPv4_ADDRESS}{The same IPv4 addressed used to start VR server}

\item{animal_class}{Character string for database}

\item{message_height}{Optional numeric value for height of question and answer boxes (default \code{1.5})}
}
\value{

}
\description{
Create a binary question scene
}
\examples{
\donttest{
animals <- binary_question_scene("Do you see any koalas in this image?", "Yes", "No", img_paths)
}

}
