% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_check.R
\name{check}
\alias{check}
\title{Check if user annotations are correct or not}
\usage{
check(img_number)
}
\arguments{
\item{img_number}{Integer to indicate which image to check as per the order the user annotated the images in}
}
\description{
Check if user annotations are correct or not
}
\examples{
\donttest{
check(1)
check(2)
check(3)
}

}
