% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_euclidean_distance.R
\name{euclidean_distance}
\alias{euclidean_distance}
\title{Calculates and returns the Euclidean Distance between two Points}
\usage{
euclidean_distance(p1, p2)
}
\arguments{
\item{p1}{First Point of type Point2d || Point3d}

\item{p2}{First Point of type Point2d || Point3d}
}
\value{
The distance between p1 and p2
}
\description{
Calculates and returns the Euclidean Distance between two Points
}
\examples{
\donttest{
euclidean_distance(p1, p2)
}

}
