% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_fixed_points.R
\name{fixed_markers}
\alias{fixed_markers}
\title{Generate markers in fixed locations}
\usage{
fixed_markers(
  module = MODULE,
  image_paths_and_points = selected_image_paths_and_points,
  marker_depth = NA
)
}
\arguments{
\item{module}{String to indicate if "2d" or "3d"}

\item{image_paths_and_points}{List of Lists containing the 'img' and the 'img_points' for the annotation markers}

\item{marker_depth}{Optional integer for the '3d' module to alter the depth of the marker where the default is 100 e.g. A value of 80 is shallower/closer to the camera than 100}
}
\description{
Generate markers in fixed locations
}
\examples{
\donttest{
selected_img_paths_and_points <- list(
  list(img = img_paths[1], img_points = img1Points),
  list(img = img_paths[2], img_points = img1Points),
  list(img = img_paths[3], img_points = img1Points)
) 

fixed_markers()
fixed_markers("2d")
fixed_markers("3d")
fixed_markers("3d", selected_img_paths_and_points, 90)              
                                                   
}

}
