% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_generate_evaluation_questions.R
\name{generate_evaluation_questions}
\alias{generate_evaluation_questions}
\title{Generate the entities needed for the evaluation question, responses, and post button}
\usage{
generate_evaluation_questions(evaluation_questions = QUESTIONS_AND_RESPONSES)
}
\arguments{
\item{evaluation_questions}{list of evaluation question lists composed of a 'question', 'answerOne', 'answerTwo', 'answerThree', and a 'answerFour',}
}
\description{
Generate the entities needed for the evaluation question, responses, and post button
}
\examples{
\donttest{
evaluation_questions <- list(
list(question = "Did you enjoy this experiment?", answerOne = "Very much", answerTwo = "Yes", #' answerThree = "A little", answerFour = "No"),
list(question = "On a scale of 1-4, how would you rate your experience?", answerOne = "1",  
answerTwo = "2", answerThree = "3", answerFour = "4")

generate_evaluation_questions(evaluation_questions)
}

}
