% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{go}
\alias{go}
\title{Go to next VR scene}
\usage{
go(image_paths, index = NA, question_type = "binary")
}
\arguments{
\item{image_paths}{Character string of image paths from current working directory.}

\item{index}{Optional numeric input to select a specific image.}

\item{question_type}{Optional question type from \code{"binary"} or \code{"multivariable"}. Defaults to \code{"binary"}.}
}
\description{
Go to next VR scene
}
\examples{
\donttest{
image_paths <- c("img1.jpg", "img2.jpg", "img3.jpg")

go(image_paths, 2)
go(image_paths, 3)
}

}
