% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url.R
\name{has_url_prefix}
\alias{has_url_prefix}
\title{has_url_prefix}
\usage{
has_url_prefix(urls)
}
\arguments{
\item{urls}{a character vector of urls}
}
\value{
a logical vector the same length as url.
}
\description{
Does string have a URL prefix
}
\details{
returns true if string matches an internal set of URL prefixes. Vectorised
over URL parameter.
}
