% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_hide_markers.R
\name{hide_markers}
\alias{hide_markers}
\title{Hide all markers}
\usage{
hide_markers(number_of_markers = NUMBER_OF_MARKERS)
}
\arguments{
\item{number_of_markers}{Integer for the number of markers to hide}
}
\description{
Hide all markers
}
\examples{
\donttest{
hide_markers()
hide_markers(15)
}

}
