% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariable_question_scene.R
\name{multivariable_question_scene}
\alias{multivariable_question_scene}
\title{Create a multivariable question scene}
\usage{
multivariable_question_scene(
  the_question,
  answer_1,
  answer_2,
  answer_3,
  answer_4,
  img_paths,
  IPv4_ADDRESS,
  message_height = 1.5
)
}
\arguments{
\item{the_question}{Question message to display}

\item{answer_1}{Answer option 1}

\item{answer_2}{Answer option 2}

\item{answer_3}{Answer option 3}

\item{answer_4}{Answer option 4}

\item{img_paths}{Character string of image paths from current working directory}

\item{IPv4_ADDRESS}{The same IPv4 addressed used to start VR server}

\item{message_height}{Optional numeric value for height of question and answer boxes (default \code{1.5})}
}
\value{

}
\description{
Create a multivariable question scene
}
\examples{
\donttest{
animals <- multivariable_question_scene("Do you see any of these habitat features in this image? If you do see a feature, click on the box to select it.",
"water", "Jaguar tracks", "Scratch marks", "Dense Vegetation", img_paths)
}

}
