% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop.R
\name{pop}
\alias{pop}
\title{Pop messages in VR scene}
\usage{
pop(visible = TRUE, question_type = "binary")
}
\arguments{
\item{visible}{Optional logical to unpop or repop. Defaults to \code{TRUE}.}

\item{question_type}{Optional question type from \code{"binary"} or \code{"multivariable"}. Defaults to \code{"binary"}.}
}
\description{
Pop messages in VR scene
}
\examples{
\donttest{
## Display messages in VR scene
pop()

## Remove messages
pop(FALSE)
}

}
