% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_set_marker_and_z_properties.R
\name{set_marker_and_props}
\alias{set_marker_and_props}
\title{Set the number of markers and the size of them for the 2d or 3d module}
\usage{
set_marker_and_props(module, number_of_markers = 15, marker_size = "base")
}
\arguments{
\item{module}{String to indicate if "2d" or "3d"}

\item{number_of_markers}{Integer between 1 and 20 of the number of markers to generate}

\item{marker_size}{String to change the size of the marker, can be: "extra-small", "small", "base" (default), "large", "extra-large"}
}
\description{
Set the number of markers and the size of them for the 2d or 3d module
}
\examples{
\donttest{
set_marker_and_props("2d")
set_marker_and_props("3d", 10, "small")
set_marker_and_props("3d", 15, "large")
}

}
