% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_set_metadata.R
\name{set_metadata}
\alias{set_metadata}
\title{Set meta data values (needed for HTTP requests to appropriate table in database)}
\usage{
set_metadata(module, module_type)
}
\arguments{
\item{module}{String "2d" || "3d"}

\item{module_type}{String "training" || "testing"}
}
\description{
Set meta data values (needed for HTTP requests to appropriate table in database)
}
\examples{
\donttest{
set_metadata("2d", "training")
set_metadata("2d", "testing")
set_metadata("3d", "training")
set_metadata("3d", "testing")
}

}
