% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start.R
\name{start}
\alias{start}
\title{Start VR server}
\usage{
start(LOCAL_IP = find_IP())
}
\arguments{
\item{LOCAL_IP}{Character string of local ip address (\code{IPv4}). If unknown go to command window and enter \code{'ipconfig'}.}
}
\description{
Start VR server
}
\examples{
\donttest{
start("YOUR-LOCAL-IP")
start()
}
}
